requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
pymongo==3.12.1
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
idna==3.4
pytz==2022.5
requests==2.28.1
sqlparse==0.2.4
urllib3==1.26.12
